Adding FluentValidation for CQRS Cmds and Qrys

+++ Application/Features/TeamMgmt/Commands/Handler/DeleteTeamCommandHandler.cs
@@ -1,6 +1,6 @@
 ﻿using Application.Interfaces;
-using Application.Services;
 using Domain.Entities;
+using FluentValidation;
 using MediatR;
 
 namespace Application.Features.TeamMgmt.Commands.Handler;
@@ -9,14 +9,23 @@ public class DeleteTeamCommandHandler: IRequestHandler<DeleteTeamCommand, Team?>
 {
     
     private readonly ITeamMgmtService _teamMgmtService;
+    private readonly IValidator<DeleteTeamCommand> _validator;
     
-    public DeleteTeamCommandHandler(ITeamMgmtService teamMgmtService)
+    public DeleteTeamCommandHandler(ITeamMgmtService teamMgmtService, IValidator<DeleteTeamCommand> validator)
     {
         _teamMgmtService = teamMgmtService;
+        _validator = validator;
     }
 
     public async Task<Team?> Handle(DeleteTeamCommand request, CancellationToken cancellationToken)
     {
+        // Validate the request
+        var validationResult = await _validator.ValidateAsync(request, cancellationToken);
+        if (!validationResult.IsValid)
+        {
+            throw new ValidationException(validationResult.Errors);
+        }
+        // Delete the team
         var deletedTeam = await _teamMgmtService.DeleteTeamAsync(request.Id);
         return deletedTeam;
     }
